+++ packages/drawer/src/js/helpers/index.js
@@ -1,5 +1,7 @@
 export * from './getBreakpoint';
 export * from './getDrawer';
+export * from './getDrawerElements';
+export * from './getDrawerID';
 export * from './initialState';
 export * from './updateFocusState';
 export * from './updateGlobalState';
