+++ packages/modal/src/js/helpers/getConfig.js
@@ -1,0 +1,5 @@
+export function getConfig(el) {
+  const string = el.getAttribute(`data-${this.settings.dataConfig}`) || '';
+  const json = string.replace(/'/g, '"');
+  return (json) ? JSON.parse(json) : {};
+}
