+++ packages/modal/src/js/helpers/getModalID.js
@@ -1,0 +1,41 @@
+export function getModalID(obj) {
+  // If it's a string, return the string.
+  if (typeof obj === 'string') {
+    return obj;
+  }
+
+  // If it's an HTML element.
+  else if (typeof obj.hasAttribute === 'function') {
+    // If it's a modal open trigger, return data value.
+    if (obj.hasAttribute(`data-${this.settings.dataOpen}`)) {
+      return obj.getAttribute(`data-${this.settings.dataOpen}`);
+    }
+
+    // If it's a modal close trigger, return data value or false.
+    else if (obj.hasAttribute(`data-${this.settings.dataClose}`)) {
+      return obj.getAttribute(`data-${this.settings.dataClose}`) || false;
+    }
+
+    // If it's a modal replace trigger, return data value.
+    else if (obj.hasAttribute(`data-${this.settings.dataReplace}`)) {
+      return obj.getAttribute(`data-${this.settings.dataReplace}`);
+    }
+
+    // If it's a modal element, return the id.
+    else if (obj.closest(this.settings.selectorModal)) {
+      obj = obj.closest(this.settings.selectorModal);
+      return obj.id || false;
+    }
+
+    // Return false if no id was found.
+    else return false;
+  }
+
+  // If it has an id property, return its value.
+  else if (obj.id) {
+    return obj.id;
+  }
+
+  // Return false if no id was found.
+  else return false;
+}
